+++ src/c_examples/stepper.c
@@ -114,6 +114,14 @@ int main(int argc, char **argv)
                printf("initialisation failed (%s)\n", Io->Errr); break;}
 
     PIN_OUT(1,0,0,1) //                            initialize pin config
+    //if (pruio_gpio_config(Io, P1, PRUIO_GPIO_OUT1)) {
+                   //printf("failed setting P1 (%s)\n", Io->Errr); break;}
+    //if (pruio_gpio_config(Io, P2, PRUIO_GPIO_OUT0)) {
+                   //printf("failed setting P2 (%s)\n", Io->Errr); break;}
+    //if (pruio_gpio_config(Io, P3, PRUIO_GPIO_OUT0)) {
+                   //printf("failed setting P3 (%s)\n", Io->Errr); break;}
+    //if (pruio_gpio_config(Io, P4, PRUIO_GPIO_OUT1)) {
+                   //printf("failed setting P4 (%s)\n", Io->Errr); break;}
 
     //' pin config OK, transfer local settings to PRU and start PRU driver
     if (pruio_config(Io, 1, 0x1FE, 0, 4)) {