+++ src/c_wrapper/pruio.h
@@ -31,36 +31,36 @@ Copyright 2014 by \Email
 //! version string
 #define PRUIO_VERSION "0.2"
 
-typedef signed char int8;      //!< 8 bit signed integer data type
-typedef short int16;           //!< 16 bit signed integer data type
-typedef int int32;             //!< 32 bit signed integer data type
-typedef unsigned char uint8;   //!< 8 bit unsigned integer data type
-typedef unsigned short uint16; //!< 16 bit unsigned integer data type
-typedef unsigned int uint32;   //!< 32 bit unsigned integer data type
-typedef float float_t;         //!< float data type
+typedef signed char int8;      //!< 8 bit signed integer data type.
+typedef short int16;           //!< 16 bit signed integer data type.
+typedef int int32;             //!< 32 bit signed integer data type.
+typedef unsigned char uint8;   //!< 8 bit unsigned integer data type.
+typedef unsigned short uint16; //!< 16 bit unsigned integer data type.
+typedef unsigned int uint32;   //!< 32 bit unsigned integer data type.
+typedef float float_t;         //!< float data type.
 
-//! tell pruss_intc_mapping.bi that we use ARM33xx
+//! Tell pruss_intc_mapping.bi that we use ARM33xx.
 #define AM33XX
 
 //! forward declaration
 typedef struct pruIo pruIo;
 
 //#include "pruio_adc.h"
-//! the default setting for avaraging
+//! The default setting for avaraging.
 #define PRUIO_DEF_AVRAGE 4
-//! the default value for open delay in channel settings
+//! The default value for open delay in channel settings.
 #define PRUIO_DEF_ODELAY 183
-//! the default value for sample delay in channel settings
+//! The default value for sample delay in channel settings.
 #define PRUIO_DEF_SDELAY 0
-//! the default number of samples to use (configures single mode)
+//! The default number of samples to use (configures single mode).
 #define PRUIO_DEF_SAMPLS 1
-//! the default step mask (steps 1 to 8 for AIN0 to AIN7, no charge step)
+//! The default step mask (steps 1 to 8 for AIN0 to AIN7, no charge step).
 #define PRUIO_DEF_STPMSK 510 // &b111111110
-//! the default timer value (sampling rate)
+//! The default timer value (sampling rate).
 #define PRUIO_DEF_TIMERV 0
-//! the default bit mode (4 = 16 bit encoding)
+//! The default bit mode (4 = 16 bit encoding).
 #define PRUIO_DEF_LSLMOD 4
-//! the default clock divisor (0 = full speed AFE = 2.4 MHz)
+//! The default clock divisor (0 = full speed AFE = 2.4 MHz).
 #define PRUIO_DEF_CLKDIV 0
 
 /** \brief Wrapper structure for AdcSteps.
@@ -69,8 +69,8 @@ typedef struct pruIo pruIo;
 */
 struct adcSteps{
   uint32
-Confg,//!< context for configuration register
-Delay;//!< context for delay register
+    Confg,//!< Context for configuration register.
+    Delay;//!< Context for delay register.
 };
 
 /** \brief Wrapper structure for AdcSet.
@@ -79,38 +79,38 @@ Delay;//!< context for delay register
 */
 typedef struct adcSet{
   uint32
-DeAd,  //!< device address
-ClAd,  //!< clock address
-ClVa;  //!< clock value
+    DeAd,  //!< Subsystem address.
+    ClAd,  //!< Clock address.
+    ClVa;  //!< Clock value.
 
   uint32
-REVISION,      //!< Register at offset 00h (chap. 12.5.1.1).
-SYSCONFIG,     //!< Register at offset 10h (chap. 12.5.1.2).
-IRQSTATUS_RAW, //!< Register at offset 24h (chap. 12.5.1.3).
-IRQSTATUS,     //!< Register at offset 28h (chap. 12.5.1.4).
-IRQENABLE_SET, //!< Register at offset 2Ch (chap. 12.5.1.5).
-IRQENABLE_CLR, //!< Register at offset 30h (chap. 12.5.1.6).
-IRQWAKEUP,     //!< Register at offset 34h (chap. 12.5.1.7).
-DMAENABLE_SET, //!< Register at offset 38h (chap. 12.5.1.8).
-DMAENABLE_CLR, //!< Register at offset 3Ch (chap. 12.5.1.9).
-CTRL,          //!< Register at offset 40h (chap. 12.5.1.10).
-ADCSTAT,       //!< Register at offset 44h (chap. 12.5.1.11).
-ADCRANGE,      //!< Register at offset 48h (chap. 12.5.1.12).
-ADC_CLKDIV,    //!< Register at offset 4Ch (chap. 12.5.1.13).
-ADC_MISC,      //!< Register at offset 50h (chap. 12.5.1.14).
-STEPENABLE,    //!< Register at offset 54h (chap. 12.5.1.15).
-IDLECONFIG;    //!< Register at offset 58h (chap. 12.5.1.16).
+    REVISION,      //!< Register at offset 00h (chap. 12.5.1.1).
+    SYSCONFIG,     //!< Register at offset 10h (chap. 12.5.1.2).
+    IRQSTATUS_RAW, //!< Register at offset 24h (chap. 12.5.1.3).
+    IRQSTATUS,     //!< Register at offset 28h (chap. 12.5.1.4).
+    IRQENABLE_SET, //!< Register at offset 2Ch (chap. 12.5.1.5).
+    IRQENABLE_CLR, //!< Register at offset 30h (chap. 12.5.1.6).
+    IRQWAKEUP,     //!< Register at offset 34h (chap. 12.5.1.7).
+    DMAENABLE_SET, //!< Register at offset 38h (chap. 12.5.1.8).
+    DMAENABLE_CLR, //!< Register at offset 3Ch (chap. 12.5.1.9).
+    CTRL,          //!< Register at offset 40h (chap. 12.5.1.10).
+    ADCSTAT,       //!< Register at offset 44h (chap. 12.5.1.11).
+    ADCRANGE,      //!< Register at offset 48h (chap. 12.5.1.12).
+    ADC_CLKDIV,    //!< Register at offset 4Ch (chap. 12.5.1.13).
+    ADC_MISC,      //!< Register at offset 50h (chap. 12.5.1.14).
+    STEPENABLE,    //!< Register at offset 54h (chap. 12.5.1.15).
+    IDLECONFIG;    //!< Register at offset 58h (chap. 12.5.1.16).
 
 //! step configuration (chap. 12.5.1.16 ff, charge step + 16 steps, by default steps 1 to 8 are used for AIN0 to AIN7).
   struct adcSteps St_p[16 + 1];
 
   uint32
-FIFO0COUNT,    //!< Register at offset E4h (chap. 12.5.1.51).
-FIFO0THRESHOLD,//!< Register at offset E8h (chap. 12.5.1.52).
-DMA0REQ,       //!< Register at offset ECh (chap. 12.5.1.53).
-FIFO1COUNT,    //!< Register at offset F0h (chap. 12.5.1.54).
-FIFO1THRESHOLD,//!< Register at offset F4h (chap. 12.5.1.55).
-DMA1REQ;       //!< Register at offset F8h (chap. 12.5.1.56).
+    FIFO0COUNT,    //!< Register at offset E4h (chap. 12.5.1.51).
+    FIFO0THRESHOLD,//!< Register at offset E8h (chap. 12.5.1.52).
+    DMA0REQ,       //!< Register at offset ECh (chap. 12.5.1.53).
+    FIFO1COUNT,    //!< Register at offset F0h (chap. 12.5.1.54).
+    FIFO1THRESHOLD,//!< Register at offset F4h (chap. 12.5.1.55).
+    DMA1REQ;       //!< Register at offset F8h (chap. 12.5.1.56).
 } adcSet;
 
 /** \brief Wrapper structure for AdcUdt.
@@ -118,19 +118,19 @@ DMA1REQ;       //!< Register at offset F8h (chap. 12.5.1.56).
 \since 0.2
 */
 typedef struct adcUdt{
-  pruIo* Top; //!< pointer to the calling PruIo instance
+  pruIo* Top; //!< Pointer to the calling PruIo instance.
   adcSet
-*Init,         //!< initial device configuration, used in the destructor  PruIo:~PruIo()
-*Conf;         //!< current device configuration, used in  PruIo::config()
+    *Init,    //!< Initial subsystem configuration, used in the destructor  PruIo:~PruIo().
+    *Conf;    //!< Current subsystem configuration, used in  PruIo::config().
   uint32
-Samples,       //!< number of samples (specifies run mode: 0 = config, 1 = IO mode, >1 = MM mode)
-TimerVal,      //!< timer value in [ns]
-InitParA;      //!< offset to read data block offset
+    Samples,  //!< Number of samples (specifies run mode: 0 = config, 1 = IO mode, >1 = MM mode).
+    TimerVal, //!< Timer value in [ns].
+    InitParA; //!< Offset to read data block offset.
   uint16
-LslMode,       //!< bit shift modus (0 to 4, for 12 to 16 bits)
-ChAz;          //!< the number of active steps
+    LslMode,  //!< Bit shift modus (0 to 4, for 12 to 16 bits).
+    ChAz;     //!< The number of active steps.
   uint16
-*Value;        //!< fetched ADC samples
+    *Value;   //!< Fetched ADC samples.
 } adcUdt;
 
 //#include "pruio_gpio.h"
@@ -140,37 +140,37 @@ ChAz;          //!< the number of active steps
 */
 typedef struct gpioSet{
   uint32
-DeAd,            //!< device address
-ClAd,            //!< clock address
-ClVa;            //!< clock value
+    DeAd,            //!< Subsystem address.
+    ClAd,            //!< Clock address.
+    ClVa;            //!< Clock value.
 
   uint32
-REVISION,        //!< Register at offset 00h (chap. 25.4.1.1).
-SYSCONFIG,       //!< Register at offset 10h (chap. 25.4.1.2).
-EOI,             //!< Register at offset 20h (chap. 25.4.1.3).
-IRQSTATUS_RAW_0, //!< Register at offset 24h (chap. 25.4.1.4).
-IRQSTATUS_RAW_1, //!< Register at offset 28h (chap. 25.4.1.5).
-IRQSTATUS_0,     //!< Register at offset 2Ch (chap. 25.4.1.6).
-IRQSTATUS_1,     //!< Register at offset 30h (chap. 25.4.1.7).
-IRQSTATUS_SET_0, //!< Register at offset 34h (chap. 25.4.1.8).
-IRQSTATUS_SET_1, //!< Register at offset 38h (chap. 25.4.1.9).
-IRQSTATUS_CLR_0, //!< Register at offset 3Ch (chap. 25.4.1.10).
-IRQSTATUS_CLR_1, //!< Register at offset 40h (chap. 25.4.1.11).
-IRQWAKEN_0,      //!< Register at offset 44h (chap. 25.4.1.12).
-IRQWAKEN_1,      //!< Register at offset 48h (chap. 25.4.1.13).
-SYSSTATUS,       //!< Register at offset 114h (chap. 25.4.1.14).
-CTRL,            //!< Register at offset 130h (chap. 25.4.1.15).
-OE,              //!< Register at offset 134h (chap. 25.4.1.16).
-DATAIN,          //!< Register at offset 138h (chap. 25.4.1.17).
-DATAOUT,         //!< Register at offset 13Ch (chap. 25.4.1.18).
-LEVELDETECT0,    //!< Register at offset 140h (chap. 25.4.1.19).
-LEVELDETECT1,    //!< Register at offset 144h (chap. 25.4.1.20).
-RISINGDETECT,    //!< Register at offset 148h (chap. 25.4.1.21).
-FALLINGDETECT,   //!< Register at offset 14Ch (chap. 25.4.1.22).
-DEBOUNCENABLE,   //!< Register at offset 150h (chap. 25.4.1.23).
-DEBOUNCINGTIME,  //!< Register at offset 154h (chap. 25.4.1.24).
-CLEARDATAOUT,    //!< Register at offset 190h (chap. 25.4.1.25).
-SETDATAOUT;      //!< Register at offset 194h (chap. 25.4.1.26).
+    REVISION,        //!< Register at offset 00h (chap. 25.4.1.1).
+    SYSCONFIG,       //!< Register at offset 10h (chap. 25.4.1.2).
+    EOI,             //!< Register at offset 20h (chap. 25.4.1.3).
+    IRQSTATUS_RAW_0, //!< Register at offset 24h (chap. 25.4.1.4).
+    IRQSTATUS_RAW_1, //!< Register at offset 28h (chap. 25.4.1.5).
+    IRQSTATUS_0,     //!< Register at offset 2Ch (chap. 25.4.1.6).
+    IRQSTATUS_1,     //!< Register at offset 30h (chap. 25.4.1.7).
+    IRQSTATUS_SET_0, //!< Register at offset 34h (chap. 25.4.1.8).
+    IRQSTATUS_SET_1, //!< Register at offset 38h (chap. 25.4.1.9).
+    IRQSTATUS_CLR_0, //!< Register at offset 3Ch (chap. 25.4.1.10).
+    IRQSTATUS_CLR_1, //!< Register at offset 40h (chap. 25.4.1.11).
+    IRQWAKEN_0,      //!< Register at offset 44h (chap. 25.4.1.12).
+    IRQWAKEN_1,      //!< Register at offset 48h (chap. 25.4.1.13).
+    SYSSTATUS,       //!< Register at offset 114h (chap. 25.4.1.14).
+    CTRL,            //!< Register at offset 130h (chap. 25.4.1.15).
+    OE,              //!< Register at offset 134h (chap. 25.4.1.16).
+    DATAIN,          //!< Register at offset 138h (chap. 25.4.1.17).
+    DATAOUT,         //!< Register at offset 13Ch (chap. 25.4.1.18).
+    LEVELDETECT0,    //!< Register at offset 140h (chap. 25.4.1.19).
+    LEVELDETECT1,    //!< Register at offset 144h (chap. 25.4.1.20).
+    RISINGDETECT,    //!< Register at offset 148h (chap. 25.4.1.21).
+    FALLINGDETECT,   //!< Register at offset 14Ch (chap. 25.4.1.22).
+    DEBOUNCENABLE,   //!< Register at offset 150h (chap. 25.4.1.23).
+    DEBOUNCINGTIME,  //!< Register at offset 154h (chap. 25.4.1.24).
+    CLEARDATAOUT,    //!< Register at offset 190h (chap. 25.4.1.25).
+    SETDATAOUT;      //!< Register at offset 194h (chap. 25.4.1.26).
 } gpioSet;
 
 /** \brief Wrapper structure for GpioArr.
@@ -179,11 +179,11 @@ SETDATAOUT;      //!< Register at offset 194h (chap. 25.4.1.26).
 */
 typedef struct gpioArr{
   uint32
-DeAd;    //!< base address of GPIO device + 0x100
+    DeAd;    //!< Base address of GPIO subsystem + 0x100.
   uint32
-DATAIN,  //!< current Value of DATAIN register (IO)
-DATAOUT, //!< current Value of DATAOUT register (IO)
-Mix;     //!< current state of pins (IN&OUT mixed)
+    DATAIN,  //!< Current Value of DATAIN register (IO).
+    DATAOUT, //!< Current Value of DATAOUT register (IO).
+    Mix;     //!< Current state of pins (IN&OUT mixed).
 } gpioArr;
 
 
@@ -194,14 +194,76 @@ Mix;     //!< current state of pins (IN&OUT mixed)
 typedef struct gpioUdt{
   pruIo* Top;//!< pointer to the calling PruIo instance
   gpioSet
-*Init[PRUIO_AZ_GPIO + 1], //!< initial device configuration, used in the destructor  PruIo:~PruIo()
-*Conf[PRUIO_AZ_GPIO + 1]; //!< current device configuration, used in  PruIo::config()
+    *Init[PRUIO_AZ_GPIO + 1], //!< Initial subsystem configuration, used in the destructor  PruIo:~PruIo().
+    *Conf[PRUIO_AZ_GPIO + 1]; //!< Current subsystem configuration, used in  PruIo::config().
   gpioArr
-*Raw[PRUIO_AZ_GPIO + 1];  //!< pointer to current raw device data (IO), all 32 bits
-  uint32 InitParA;       //!< offset to read data block offset
+    *Raw[PRUIO_AZ_GPIO + 1];  //!< Pointer to current raw subsystem data (IO), all 32 bits.
+  uint32 InitParA;            //!< Offset to read data block offset.
 } gpioUdt;
 
 
+
+
+//#include "pruio_timer.bi"
+/** \brief Wrapper structure for TimerSet.
+
+\since 0.4
+*/
+typedef struct timerSet{
+  uint32
+    DeAd, //!< Subsystem address.
+    ClAd, //!< Clock address.
+    ClVa; //!< Clock value.
+
+  uint32
+    TIDR         //!< Register at offset  00h (see \ArmRef{20.1.5.1} ).
+  , TIOCPCFG     //!< Timer OCP Configuration Register (offset 10h, see \ArmRef{20.1.5.2} ).
+  , IRQEOI       //!< Timer IRQ End-of-Interrupt Register (offset 20h, see \ArmRef{20.1.5.3} ).
+  , IRQSTATUSRAW //!< Timer Status Raw Register (offset 24h, see \ArmRef{20.1.5.4} ).
+  , IRQSTATUS    //!< Timer Status Register (offset 28h, see \ArmRef{20.1.5.5} ).
+  , IRQENABLESET //!< Timer Interrupt Enable Set Register (offset 2Ch, see \ArmRef{20.1.5.6} ).
+  , IRQENABLECLR //!< Timer Interrupt Enable Clear Register (offset 30h, see \ArmRef{20.1.5.7} ).
+  , IRQWAKEEN    //!< Timer IRQ Wakeup Enable Register (offset 34h, see \ArmRef{20.1.5.8} ).
+  , TCLR   //!< Timer Control Register (offset 38h, see \ArmRef{20.1.5.9} ).
+  , TCRR   //!< Timer Counter Register (offset 3Ch, see \ArmRef{20.1.5.10} ).
+  , TLDR   //!< Timer Load Register (offset 40h, see \ArmRef{20.1.5.11} ).
+  , TTGR   //!< Timer Trigger Register (offset 44h, see \ArmRef{20.1.5.12} ).
+  , TWPS   //!< Timer Write Posting Bits Register (offset 48h, see \ArmRef{20.1.5.13} ).
+  , TMAR   //!< Timer Match Register (offset 4Ch, see \ArmRef{20.1.5.14} ).
+  , TCAR1  //!< Timer Capture Register (offset 50h, see \ArmRef{20.1.5.15} ).
+  , TSICR  //!< Timer Synchronous Interface Control Register (offset 54h, see \ArmRef{20.1.5.16} ).
+  , TCAR2; //!< Timer Capture Register (offset 58h, see \ArmRef{20.1.5.17} ).
+} timerSet;
+
+
+/** \brief Wrapper structure for TimerArr.
+
+\since 0.4
+*/
+typedef struct timerArr{
+  uint32
+    DeAd; //!< Subsystem address.
+
+  uint32
+    TCAR1 //!< Current value of TCAR1 register (IO, RB).
+  , TCAR2 //!< Current value of TCAR2 register (IO, RB).
+  , TCRR; //!< Current value of TCRR register (IO, RB).
+} timerArr;
+
+
+/** \brief Wrapper structure for TimerUdt.
+
+\since 0.4
+*/
+typedef struct timerUdt{
+  pruIo* Top;                 //!< Pointer to the calling PruIo instance.
+  timerSet
+    *Init[PRUIO_AZ_GPIO + 1], //!< Initial subsystem configuration, used in the destructor  PruIo:~PruIo().
+    *Conf[PRUIO_AZ_GPIO + 1]; //!< Current subsystem configuration, used in  PruIo::config().
+  timerArr
+    *Raw[PRUIO_AZ_GPIO + 1];  //!< Pointer to current raw subsystem data (IO), all 32 bits.
+  uint32 InitParA;            //!< Offset to read data block offset.
+} timerUdt;
 
 
 //#include "pruio_pwm.bi"
@@ -211,101 +273,101 @@ typedef struct gpioUdt{
 */
 typedef struct pwmssSet{
   uint32
-DeAd,       //!< device address
-ClAd,       //!< clock address
-ClVa;       //!< clock value
+    DeAd,       //!< Subsystem address.
+    ClAd,       //!< Clock address.
+    ClVa;       //!< Clock value.
 
   uint32
-IDVER,      //!< IP Revision Register (chap. 15.1.3.1).
-SYSCONFIG,  //!< System Configuration Register (chap. 15.1.3.2).
-CLKCONFIG,  //!< Clock Configuration Register (chap. 15.1.3.3).
-CLKSTATUS;  //!< Clock Status Register (chap. 15.1.3.4).
+    IDVER,      //!< IP Revision Register (chap. 15.1.3.1).
+    SYSCONFIG,  //!< System Configuration Register (chap. 15.1.3.2).
+    CLKCONFIG,  //!< Clock Configuration Register (chap. 15.1.3.3).
+    CLKSTATUS;  //!< Clock Status Register (chap. 15.1.3.4).
 
   uint32
-TSCTR,      //!< Time-Stamp Counter Register (chap. 15.3.4.1.1).
-CTRPHS,     //!< Counter Phase Offset Value Register (chap. 15.3.4.1.2).
-CAP1,       //!< Capture 1 Register (chap. 15.3.4.1.3).
-CAP2,       //!< Capture 2 Register (chap. 15.3.4.1.4).
-CAP3,       //!< Capture 3 Register (chap. 15.3.4.1.5).
-CAP4;       //!< Capture 4 Register (chap. 15.3.4.1.6).
+    TSCTR,      //!< Time-Stamp Counter Register (chap. 15.3.4.1.1).
+    CTRPHS,     //!< Counter Phase Offset Value Register (chap. 15.3.4.1.2).
+    CAP1,       //!< Capture 1 Register (chap. 15.3.4.1.3).
+    CAP2,       //!< Capture 2 Register (chap. 15.3.4.1.4).
+    CAP3,       //!< Capture 3 Register (chap. 15.3.4.1.5).
+    CAP4;       //!< Capture 4 Register (chap. 15.3.4.1.6).
   uint16
-ECCTL1,     //!< Capture Control Register 1 (chap. 15.3.4.1.7).
-ECCTL2,     //!< Capture Control Register 2 (chap. 15.3.4.1.8).
-ECEINT,     //!< Capture Interrupt Enable Register (chap. 15.3.4.1.9).
-ECFLG,      //!< Capture Interrupt Flag Register (chap. 15.3.4.1.10).
-ECCLR,      //!< Capture Interrupt Clear Register (chap. 15.3.4.1.11).
-ECFRC;      //!< Capture Interrupt Force Register (chap. 15.3.4.1.12).
+    ECCTL1,     //!< Capture Control Register 1 (chap. 15.3.4.1.7).
+    ECCTL2,     //!< Capture Control Register 2 (chap. 15.3.4.1.8).
+    ECEINT,     //!< Capture Interrupt Enable Register (chap. 15.3.4.1.9).
+    ECFLG,      //!< Capture Interrupt Flag Register (chap. 15.3.4.1.10).
+    ECCLR,      //!< Capture Interrupt Clear Register (chap. 15.3.4.1.11).
+    ECFRC;      //!< Capture Interrupt Force Register (chap. 15.3.4.1.12).
   uint32
-CAP_REV;    //!< Revision ID Register (chap. 15.3.4.1.13).
+    CAP_REV;    //!< Revision ID Register (chap. 15.3.4.1.13).
 
   uint32
-QPOSCNT,    //!< Position Counter Register (chap. 15.4.3.1).
-QPOSINIT,   //!< Position Counter Initialization Register (chap. 15.4.3.2).
-QPOSMAX,    //!< Maximum Position Count Register (chap. 15.4.3.3).
-QPOSCMP,    //!< Position-Compare Register 2/1 (chap. 15.4.3.4).
-QPOSILAT,   //!< Index Position Latch Register (chap. 15.4.3.5).
-QPOSSLAT,   //!< Strobe Position Latch Register (chap. 15.4.3.6).
-QPOSLAT,    //!< Position Counter Latch Register (chap. 15.4.3.7).
-QUTMR,      //!< Unit Timer Register (chap. 15.4.3.8).
-QUPRD;      //!< Unit Period Register (chap. 15.4.3.9).
+    QPOSCNT,    //!< Position Counter Register (chap. 15.4.3.1).
+    QPOSINIT,   //!< Position Counter Initialization Register (chap. 15.4.3.2).
+    QPOSMAX,    //!< Maximum Position Count Register (chap. 15.4.3.3).
+    QPOSCMP,    //!< Position-Compare Register 2/1 (chap. 15.4.3.4).
+    QPOSILAT,   //!< Index Position Latch Register (chap. 15.4.3.5).
+    QPOSSLAT,   //!< Strobe Position Latch Register (chap. 15.4.3.6).
+    QPOSLAT,    //!< Position Counter Latch Register (chap. 15.4.3.7).
+    QUTMR,      //!< Unit Timer Register (chap. 15.4.3.8).
+    QUPRD;      //!< Unit Period Register (chap. 15.4.3.9).
   uint16
-QWDTMR,     //!< Watchdog Timer Register (chap. 15.4.3.10).
-QWDPRD,     //!< Watchdog Period Register (chap. 15.4.3.11).
-QDECCTL,    //!< Decoder Control Register (chap. 15.4.3.12).
-QEPCTL,     //!< Control Register (chap. 15.4.3.14).
-QCASCTL,    //!< Capture Control Register (chap. 15.4.3.15).
-QPOSCTL,    //!< Position-Compare Control Register (chap. 15.4.3.15).
-QEINT,      //!< Interrupt Enable Register (chap. 15.4.3.16).
-QFLG,       //!< Interrupt Flag Register (chap. 15.4.3.17).
-QCLR,       //!< Interrupt Clear Register (chap. 15.4.3.18).
-QFRC,       //!< Interrupt Force Register (chap. 15.4.3.19).
-QEPSTS,     //!< Status Register (chap. 15.4.3.20).
-QCTMR,      //!< Capture Timer Register (chap. 15.4.3.21).
-QCPRD,      //!< Capture Period Register (chap. 15.4.3.22).
-QCTMRLAT,   //!< Capture Timer Latch Register (chap. 15.4.3.23).
-QCPRDLAT,   //!< Capture Period Latch Register (chap. 15.4.3.24).
-empty;      //!< adjust at uint32 border
+    QWDTMR,     //!< Watchdog Timer Register (chap. 15.4.3.10).
+    QWDPRD,     //!< Watchdog Period Register (chap. 15.4.3.11).
+    QDECCTL,    //!< Decoder Control Register (chap. 15.4.3.12).
+    QEPCTL,     //!< Control Register (chap. 15.4.3.14).
+    QCASCTL,    //!< Capture Control Register (chap. 15.4.3.15).
+    QPOSCTL,    //!< Position-Compare Control Register (chap. 15.4.3.15).
+    QEINT,      //!< Interrupt Enable Register (chap. 15.4.3.16).
+    QFLG,       //!< Interrupt Flag Register (chap. 15.4.3.17).
+    QCLR,       //!< Interrupt Clear Register (chap. 15.4.3.18).
+    QFRC,       //!< Interrupt Force Register (chap. 15.4.3.19).
+    QEPSTS,     //!< Status Register (chap. 15.4.3.20).
+    QCTMR,      //!< Capture Timer Register (chap. 15.4.3.21).
+    QCPRD,      //!< Capture Period Register (chap. 15.4.3.22).
+    QCTMRLAT,   //!< Capture Timer Latch Register (chap. 15.4.3.23).
+    QCPRDLAT,   //!< Capture Period Latch Register (chap. 15.4.3.24).
+    empty;      //!< Adjust at uint32 border.
   uint32
-QEP_REV;    //!< Revision ID (chap. 15.4.3.25
+    QEP_REV;    //!< Revision ID (chap. 15.4.3.25).
 
   uint16
-TBCTL,      //!< Time-Base Control Register
-TBSTS,      //!< Time-Base Status Register
-TBPHSHR,    //!< Extension for HRPWM Phase Register
-TBPHS,      //!< Time-Base Phase Register
-TBCNT,      //!< Time-Base Counter Register
-TBPRD,      //!< Time-Base Period Register
+    TBCTL,      //!< Time-Base Control Register.
+    TBSTS,      //!< Time-Base Status Register.
+    TBPHSHR,    //!< Extension for HRPWM Phase Register.
+    TBPHS,      //!< Time-Base Phase Register.
+    TBCNT,      //!< Time-Base Counter Register.
+    TBPRD,      //!< Time-Base Period Register.
 
-CMPCTL,     //!< Counter-Compare Control Register
-CMPAHR,     //!< Extension for HRPWM Counter-Compare A Register
-CMPA,       //!< Counter-Compare A Register
-CMPB,       //!< Counter-Compare B Register
+    CMPCTL,     //!< Counter-Compare Control Register.
+    CMPAHR,     //!< Extension for HRPWM Counter-Compare A Register.
+    CMPA,       //!< Counter-Compare A Register.
+    CMPB,       //!< Counter-Compare B Register.
 
-AQCTLA,     //!< Action-Qualifier Control Register for Output A (EPWMxA)
-AQCTLB,     //!< Action-Qualifier Control Register for Output B (EPWMxB)
-AQSFRC,     //!< Action-Qualifier Software Force Register
-AQCSFRC,    //!< Action-Qualifier Continuous S/W Force Register Set
+    AQCTLA,     //!< Action-Qualifier Control Register for Output A (EPWMxA).
+    AQCTLB,     //!< Action-Qualifier Control Register for Output B (EPWMxB).
+    AQSFRC,     //!< Action-Qualifier Software Force Register.
+    AQCSFRC,    //!< Action-Qualifier Continuous S/W Force Register Set.
 
-DBCTL,     //!< Dead-Band Generator Control Register
-DBRED,     //!< Dead-Band Generator Rising Edge Delay Count Register
-DBFED,     //!< Dead-Band Generator Falling Edge Delay Count Register
+    DBCTL,     //!< Dead-Band Generator Control Register.
+    DBRED,     //!< Dead-Band Generator Rising Edge Delay Count Register.
+    DBFED,     //!< Dead-Band Generator Falling Edge Delay Count Register.
 
-TZSEL,     //!< Trip-Zone Select Register
-TZCTL,     //!< Trip-Zone Control Register
-TZEINT,    //!< Trip-Zone Enable Interrupt Register
-TZFLG,     //!< Trip-Zone Flag Register
-TZCLR,     //!< Trip-Zone Clear Register
-TZFRC,     //!< Trip-Zone Force Register
+    TZSEL,     //!< Trip-Zone Select Register.
+    TZCTL,     //!< Trip-Zone Control Register.
+    TZEINT,    //!< Trip-Zone Enable Interrupt Register.
+    TZFLG,     //!< Trip-Zone Flag Register.
+    TZCLR,     //!< Trip-Zone Clear Register.
+    TZFRC,     //!< Trip-Zone Force Register.
 
-ETSEL,     //!< Event-Trigger Selection Register
-ETPS,      //!< Event-Trigger Pre-Scale Register
-ETFLG,     //!< Event-Trigger Flag Register
-ETCLR,     //!< Event-Trigger Clear Register
-ETFRC,     //!< Event-Trigger Force Register
+    ETSEL,     //!< Event-Trigger Selection Register.
+    ETPS,      //!< Event-Trigger Pre-Scale Register.
+    ETFLG,     //!< Event-Trigger Flag Register.
+    ETCLR,     //!< Event-Trigger Clear Register.
+    ETFRC,     //!< Event-Trigger Force Register.
 
-PCCTL,     //!< PWM-Chopper Control Register
+    PCCTL,     //!< PWM-Chopper Control Register.
 
-HRCTL;     //!< HRPWM Control Register
+    HRCTL;     //!< HRPWM Control Register.
 } pwmssSet;
 
 
@@ -315,15 +377,15 @@ HRCTL;     //!< HRPWM Control Register
 */
 typedef struct pwmssArr{
   uint32
-DeAd;      //!< device address
+    DeAd;      //!< Subsystem address.
   uint32
-CMax,      //!< maximum counter value
-C1,        //!< on time counter value
-C2,        //!< period time counter value
-fe1,       //!< ???
-fe2,       //!< ???
-fe3,       //!< ???
-fe4;       //!< ???
+    CMax,      //!< Maximum counter value.
+    C1,        //!< On time counter value.
+    C2,        //!< Period time counter value.
+    fe1,       //!< Future expansion.
+    fe2,       //!< Future expansion.
+    fe3,       //!< Future expansion.
+    fe4;       //!< Future expansion.
 } pwmssArr;
 
 
@@ -334,13 +396,13 @@ fe4;       //!< ???
 typedef struct pwmssUdt{
   pruIo* Top;  //!< pointer to the calling PruIo instance
   pwmssSet
-*Init[PRUIO_AZ_PWMSS + 1], //!< initial device configuration, used in the destructor  PruIo:~PruIo()
-*Conf[PRUIO_AZ_PWMSS + 1]; //!< current device configuration, used in  PruIo::config()
+    *Init[PRUIO_AZ_PWMSS + 1], //!< Initial subsystem configuration, used in the destructor  PruIo:~PruIo().
+    *Conf[PRUIO_AZ_PWMSS + 1]; //!< Current subsystem configuration, used in  PruIo::config().
   pwmssArr
-*Raw[PRUIO_AZ_PWMSS + 1];  //!< pointer to current raw subsystem data (IO)
-  uint32 InitParA;         //!< offset to read data block offset
+    *Raw[PRUIO_AZ_PWMSS + 1];  //!< Pointer to current raw subsystem data (IO).
+  uint32 InitParA;             //!< Offset to read data block offset.
   const uint16
-CapMod;                    //!< value for ECCTL2 in CAP mode (&b11010110)
+    CapMod;                    //!< Value for ECCTL2 in CAP mode (&b11010110).
 } pwmssUdt;
 
 
@@ -368,21 +430,42 @@ typedef struct qepMod qepMod;
 //! PRUSS driver interrupt settings
 #include "pruss_intc_mapping.h"
 
+/** \brief Wrapper enumerators for PinMuxing.
+
+\since 0.2
+*/
+enum pinMuxing{
+  PRUIO_PULL_DOWN = 0,      //!< Pulldown resistor connected (&b000000).
+  PRUIO_NO_PULL   = 1 << 3, //!< No resistor connected (&b001000).
+  PRUIO_PULL_UP   = 1 << 4, //!< Pullup resistor connected (&b010000).
+  PRUIO_RX_ACTIV  = 1 << 5, //!< Input receiver enabled (&b100000).
+  PRUIO_GPIO_OUT0 = 7 + PRUIO_NO_PULL,                    //!< GPIO output low (no resistor).
+  PRUIO_GPIO_OUT1 = 7 + PRUIO_NO_PULL + 128,              //!< GPIO output high (no resistor).
+  PRUIO_GPIO_IN   = 7 + PRUIO_NO_PULL + PRUIO_RX_ACTIV,   //!< GPIO input (no resistor).
+  PRUIO_GPIO_IN_0 = 7 + PRUIO_PULL_DOWN + PRUIO_RX_ACTIV, //!< GPIO input (pulldown resistor).
+  PRUIO_GPIO_IN_1 = 7 + PRUIO_PULL_UP + PRUIO_RX_ACTIV,   //!< GPIO input (pullup resistor).
+  PRUIO_PIN_RESET = 0xFF
+};
+
 /** \brief Wrapper enumerators for ActivateDevice.
 
 \since 0.2
 */
 enum activateDevice{
-  PRUIO_ACT_PRU1  =   1    , //!< activate PRU-1 (= default, instead of PRU-0)
-  PRUIO_ACT_ADC   = 1 << 1 , //!< activate ADC
-  PRUIO_ACT_GPIO0 = 1 << 2 , //!< activate GPIO-0
-  PRUIO_ACT_GPIO1 = 1 << 3 , //!< activate GPIO-1
-  PRUIO_ACT_GPIO2 = 1 << 4 , //!< activate GPIO-2
-  PRUIO_ACT_GPIO3 = 1 << 5 , //!< activate GPIO-3
-  PRUIO_ACT_PWM0  = 1 << 6 , //!< activate PWMSS-0 (including eCAP, eQEP, ePWM)
-  PRUIO_ACT_PWM1  = 1 << 7 , //!< activate PWMSS-1 (including eCAP, eQEP, ePWM)
-  PRUIO_ACT_PWM2  = 1 << 8 , //!< activate PWMSS-2 (including eCAP, eQEP, ePWM)
-  PRUIO_DEF_ACTIVE = 0xFFFF  //!< activate all devices
+  PRUIO_ACT_PRU1  =   1    , //!< Activate PRU-1 (= default, instead of PRU-0).
+  PRUIO_ACT_ADC   = 1 << 1 , //!< Activate ADC.
+  PRUIO_ACT_GPIO0 = 1 << 2 , //!< Activate GPIO-0.
+  PRUIO_ACT_GPIO1 = 1 << 3 , //!< Activate GPIO-1.
+  PRUIO_ACT_GPIO2 = 1 << 4 , //!< Activate GPIO-2.
+  PRUIO_ACT_GPIO3 = 1 << 5 , //!< Activate GPIO-3.
+  PRUIO_ACT_PWM0  = 1 << 6 , //!< Activate PWMSS-0 (including eCAP, eQEP, ePWM).
+  PRUIO_ACT_PWM1  = 1 << 7 , //!< Activate PWMSS-1 (including eCAP, eQEP, ePWM).
+  PRUIO_ACT_PWM2  = 1 << 8 , //!< Activate PWMSS-2 (including eCAP, eQEP, ePWM).
+  PRUIO_ACT_TIM4  = 1 << 9 , //!< Activate TIMER-4.
+  PRUIO_ACT_TIM5  = 1 << 10, //!< Activate TIMER-5.
+  PRUIO_ACT_TIM6  = 1 << 11, //!< Activate TIMER-6.
+  PRUIO_ACT_TIM7  = 1 << 12, //!< Activate TIMER-7.
+  PRUIO_DEF_ACTIVE = 0xFFFFF //!< Activate all subsystems.
 };
 
 
@@ -392,7 +475,7 @@ enum activateDevice{
 */
 typedef struct ballSet{
   uint32
-DeAd;//!< base address of Control Module device
+    DeAd;  //!< Base address of Control Module subsystem.
   uint8 Value[PRUIO_AZ_BALL + 1];//!< The values of the pad control registers.
 } ballSet;
 
@@ -401,43 +484,42 @@ DeAd;//!< base address of Control Module device
 \since 0.0
 */
 typedef struct pruIo{
-  adcUdt* Adc;//!< pointer to ADC device structure
-  gpioUdt* Gpio;//!< pointer to GPIO device structure
-  pwmssUdt* PwmSS;//!< pointer to PWMSS device structure
-  pwmMod* Pwm;//!< pointer to the ePWM module structure (in PWMSS devices)
-  capMod* Cap;//!< pointer to the eCAP module structure (in PWMSS devices)
-  //qepMod* Qep;//!< pointer to the eQEP module structure (in PWMSS devices)
+  adcUdt* Adc;     //!< Pointer to ADC subsystem structure.
+  gpioUdt* Gpio;   //!< Pointer to GPIO subsystem structure.
+  pwmssUdt* PwmSS; //!< Pointer to PWMSS subsystem structure.
+  timerUdt* TimSS; //!< Pointer to TIMER subsystem structure.
+  pwmMod* Pwm;     //!< Pointer to the ePWM module structure (in PWMSS subsystems).
+  capMod* Cap;     //!< Pointer to the eCAP module structure (in PWMSS subsystems).
+  //qepMod* Qep;//!< pointer to the eQEP module structure (in PWMSS subsystems)
 
-  char* Errr;//!< pointer for error messages
-  uint32* DRam;//!< pointer to access PRU DRam
+  char* Errr;      //!< Pointer for error messages.
+  uint32* DRam;    //!< Pointer to access PRU DRam.
   ballSet
-*Init,//!< The devices register data at start-up (to restore devices at the end)
-*Conf;//!< The devices register data used by libpruio (current local data)
+    *Init,         //!< The subsystems register data at start-up (to restore subsystems at the end).
+    *Conf;         //!< The subsystems register data used by libpruio (current local data).
   void
-*ERam,//!< pointer to read PRU external ram
-*DInit,//!< pointer to block of devices initial data
-*DConf,//!< pointer to block of devices configuration data
-*MOffs;//!< configuration offset for modules
+    *ERam,         //!< Pointer to read PRU external ram.
+    *DInit,        //!< Pointer to block of subsystems initial data.
+    *DConf,        //!< Pointer to block of subsystems configuration data.
+    *MOffs;        //!< Configuration offset for modules.
   uint8
-*BallOrg,//!< pointer for original Ball configuration
-*BallConf;//!< pointer to ball configuration (CPU pin muxing)
+    *BallOrg,      //!< Pointer for original Ball configuration.
+    *BallConf;     //!< Pointer to ball configuration (CPU pin muxing).
   uint32
-EAddr,//!< the address of the external memory (PRUSS-DDR)
-ESize,//!< the size of the external memory (PRUSS-DDR)
-DSize,//!< the size of a data block (DInit or DConf)
-PruNo,//!< the PRU number to use (defaults to 1)
-PruEvtOut,//!< the interrupt channel to send commands to PRU
-PruIRam,//!< the PRU instruction ram to load
-PruDRam;//!< the PRU data ram
+    EAddr,         //!< The address of the external memory (PRUSS-DDR).
+    ESize,         //!< The size of the external memory (PRUSS-DDR).
+    DSize,         //!< The size of a data block (DInit or DConf).
+    PruNo,         //!< The PRU number to use (defaults to 1).
+    PruIRam,       //!< The PRU instruction ram to load.
+    PruDRam;       //!< The PRU data ram.
   int16
-ArmPruInt,//!< the interrupt to send
-ParOffs,//!< the offset for the parameters of a module
-DevAct;//!< active devices
+    ParOffs,       //!< The offset for the parameters of a module.
+    DevAct;        //!< Active subsystems.
 
-//! interrupt settings (we also set default interrupts, so that the other PRUSS can be used in parallel)
+//! Interrupt settings (we also set default interrupts, so that the other PRUSS can be used in parallel).
   struct __pruss_intc_initdata IntcInit;
 
-//! list of GPIO numbers, corresponding to ball index
+//! List of GPIO numbers, corresponding to ball index.
   uint8 BallGpio[PRUIO_AZ_BALL + 1];
 } pruIo;
 
@@ -449,9 +531,9 @@ DevAct;//!< active devices
 \param SaD sample delay for default steps (0 to 255, defaults to 0)
 \returns A pointer for the new instance.
 
-Since the constructor reads the original devices configurations and the
-destructor restores them, it's recommended to create and use just one
-PruIo instance at the same time.
+Since the constructor reads the original subsystem configurations and
+the destructor restores them, it's recommended to create and use just
+one PruIo instance at the same time.
 
 \since 0.0
 */
@@ -514,7 +596,7 @@ char* pruio_rb_start(pruIo* Io);
 
 \since 0.2
 */
-char* pruio_gpio_setValue(pruIo* Io, uint8 Ball, uint8 Modus);
+char* pruio_gpio_config(pruIo* Io, uint8 Ball, uint8 Modus);
 
 /** \brief Wrapper function for GpioUdt::setValue().
 \param Io The pointer of the  PruIo instance